Refactoring UI Testing Toolbox tests to be in their own project
